+++ apps/researcher/src/app/[locale]/(objects)/search-field.tsx
@@ -3,24 +3,22 @@
 import {useListStore} from '@colonial-collections/list-store';
 import {MagnifyingGlassIcon} from '@heroicons/react/24/solid';
 import {useTranslations} from 'next-intl';
-import {useEffect, useState} from 'react';
+import {useRouter} from 'next-intl/client';
+import {useState} from 'react';
 
 export function SearchField({placeholder = ''}: {placeholder?: string}) {
   const query = useListStore(s => s.query);
   const [inputText, setInputText] = useState(query);
-  const queryChange = useListStore(s => s.queryChange);
   const t = useTranslations('Filters');
-
-  useEffect(() => {
-    setInputText(query);
-  }, [query]);
+  const router = useRouter();
 
   const handleQueryChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setInputText(e.target.value);
   };
 
   const search = () => {
-    queryChange(inputText);
+    const urlSearchParams = new URLSearchParams({query: inputText});
+    router.replace(`/?${urlSearchParams}`);
   };
 
   const handleKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
@@ -30,7 +28,7 @@ export function SearchField({placeholder = ''}: {placeholder?: string}) {
   };
 
   return (
-    <div className="flex flex-row w-full">
+    <div className="w-full flex justify-between">
       <input
         data-testid="searchQuery"
         value={inputText}
@@ -38,37 +36,18 @@ export function SearchField({placeholder = ''}: {placeholder?: string}) {
         type="text"
         name="search"
         id="search"
-        className="py-1 px-3 w-full rounded-l border border-neutral-700"
+        className="p-3 rounded-l placeholder:text-blueGrey-500 placeholder:italic w-full text-consortiumBlueOld-800 not-italic"
         aria-label={t('accessibilityTypeToFilter')}
         placeholder={placeholder}
         onKeyUp={handleKeyPress}
       />
       <button
-        className="bg-neutral-700 py-1 px-3 rounded-r border-t  border-b border-r border-neutral-700"
+        className="flex items-center p-3 rounded-r bg-consortiumGreen-300"
         aria-label={t('accessibilityTypeToFilter')}
         onClick={search}
       >
-        <MagnifyingGlassIcon className="w-4 h-4 fill-white" />
+        <MagnifyingGlassIcon className="w-6 h-6 fill-consortiumBlue-800" />
       </button>
-    </div>
-  );
-}
-
-function Label() {
-  const t = useTranslations('Filters');
-
-  return (
-    <label htmlFor="search" className="font-semibold">
-      {t('search')}
-    </label>
-  );
-}
-
-export function SearchFieldWithLabel() {
-  return (
-    <div className="w-full max-w-[450px] relative" id="facets">
-      <Label />
-      <SearchField />
     </div>
   );
 }
