+++ packages/ui/branding/background.tsx
@@ -1,0 +1,38 @@
+export function Background() {
+  return (
+    <svg
+      version="1.1"
+      fill="currentColor"
+      className="w-full h-auto fill-consortiumGreen-500"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      x="0px"
+      y="0px"
+      viewBox="0 0 1474 314"
+      xmlSpace="preserve"
+    >
+      <g>
+        <polygon points="1190.2,314 1415.8,137.5 1426,314 1427.8,314 1417.5,133.9 1187.2,314 	"></polygon>
+        <path
+          d="M0,174.7l56.7-34.6L26.4,314h1.9l30.5-175.2L122,96.8L220.3,246l1.2,68h1.8l-1.2-67.2l90.7,13.7l-7.5,53.5h1.9
+			l31.5-224.9l192.2,7.1l-14.3,186.8l0,0.3l20.7,30.7h2.2l-11-16.3l71.5-224.2l23.5,4.8L548.3,314h1.9l75-235.4l23.5,4.8L575.2,314
+			h1.9l73.4-230.3l23.5,4.8L602.1,314h1.9l71.8-225.2l15,3.1l0.5,0.1l13.2-15L629,314h1.9l76.7-240.5l39.2-44.6L656,314h1.9l92-288.6
+			l18.5-21l12,3.8L682.9,314h1.9L782.1,8.8l22.7,7.3L709.8,314h1.9l94.8-297.3l22.7,7.3l-92.4,290h1.9L831,24.5l22.7,7.3l-90,282.2
+			h1.9l89.8-281.7l22.7,7.3L790.6,314h1.9l87.3-273.8l22.7,7.3L823,296.9l-18.4,11.3l-0.3,0.2l-2,5.7h2l1.6-4.4l51.4-31.5l0.5,0.2
+			l0.2-0.6l123.5-75.7l72.4-14l-9.2,126.1h1.8l9.3-127.2l0.1-1.2L982,200l-5.4-41l0.2-0.5l126.5-38.6l15,194.1h1.8l-15.2-196.5
+			l-99.8,30.4l-16.2-27.5l0.4-1.3l-1.5-0.5l-33-56l-0.2-0.3L768.3,2.5l-0.6-0.2l-77,87.7L295.4,9.8l-74.5,233.8L123,95l-0.5-0.8
+			l-65.1,43.3L0,172.6V174.7z M980.2,200.6l-20.9,12.8l16-50.3L980.2,200.6z M988.1,122.7l15.2,25.7l-25.9,7.9L988.1,122.7z
+			 M986.9,120.6L975.3,157l-0.8,0.2l0.2,1.6l-17.9,56.1l-31.1,19l44.9-140.8L986.9,120.6z M953.4,63.8l16,27.1l-46,144.4l-31.1,19
+			l60.8-190.7L953.4,63.8z M951.3,63.1L890.2,255l0.8,0.2l-32.2,19.7l69.8-219.1L951.3,63.1z M926.9,55.3l-70.5,221.1l-31.1,19
+			L904.2,48L926.9,55.3z M313,258.7l-13.8-2.1l25.2-79L313,258.7z M598.1,73.1l-71,222.6l-8.7-13l3.8-50.1l52.4-164.3L598.1,73.1z
+			 M572.8,67.9l-49.9,156.7l7.9-103.2l18.6-58.3L572.8,67.9z M547.5,62.8l-16.2,50.7l1.5-19.1l-20.1-0.7l11.3-35.6L547.5,62.8z
+			 M522.2,57.7l-11.4,35.9l-24.7-0.9l12.7-39.7L522.2,57.7z M497,52.6l-12.8,40l-24.7-0.9l14-43.9L497,52.6z M471.6,47.4l-14.1,44.2
+			l-24.7-0.9l15.3-48L471.6,47.4z M446.4,42.3L431,90.6l-24.7-0.9l16.6-52.2L446.4,42.3z M421.1,37.2l-16.7,52.5l-24.7-0.9l18-56.3
+			L421.1,37.2z M395.8,32.1l-18,56.6L353,87.8l19.3-60.5L395.8,32.1z M370.5,26.9l-19.4,60.8L337,87.2l-11.1,79.6l-28.5,89.5
+			l-23.9-3.6L347,22.2L370.5,26.9z M345.2,21.8l-73.5,230.6l-23.9-3.6L321.7,17L345.2,21.8z M296.6,12l23.3,4.7L246,248.5l-23.6-3.6
+			L296.6,12z"
+        ></path>
+      </g>
+    </svg>
+  );
+}
