+++ packages/ui/branding/index.tsx
@@ -1,1 +1,2 @@
 export * from './consortium-logo';
+export * from './background';
